Move request resolvers from Pimcore\Service to Pimcore\Http\Request\Resolver namespace
