+++ pimcore/lib/Pimcore/Bundle/AdminBundle/EventListener/BruteforceProtectionListener.php
@@ -18,7 +18,7 @@ use Pimcore\Bundle\AdminBundle\Controller\BruteforceProtectedControllerInterface
 use Pimcore\Bundle\AdminBundle\Security\BruteforceProtectionHandler;
 use Pimcore\Bundle\AdminBundle\Security\Exception\BruteforceProtectionException;
 use Pimcore\Bundle\CoreBundle\EventListener\Traits\PimcoreContextAwareTrait;
-use Pimcore\Service\Request\PimcoreContextResolver;
+use Pimcore\Http\Request\Resolver\PimcoreContextResolver;
 use Symfony\Component\EventDispatcher\EventSubscriberInterface;
 use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\HttpKernel\Event\FilterControllerEvent;
