+++ pimcore/lib/Pimcore/Bundle/CoreBundle/DependencyInjection/Compiler/PimcoreContextResolverAwarePass.php
@@ -15,7 +15,7 @@
 namespace Pimcore\Bundle\CoreBundle\DependencyInjection\Compiler;
 
 use Pimcore\Http\Context\PimcoreContextResolverAwareInterface;
-use Pimcore\Service\Request\PimcoreContextResolver;
+use Pimcore\Http\Request\Resolver\PimcoreContextResolver;
 use Symfony\Component\DependencyInjection\Compiler\CompilerPassInterface;
 use Symfony\Component\DependencyInjection\ContainerBuilder;
 use Symfony\Component\DependencyInjection\Reference;
