+++ pimcore/lib/Pimcore/Bundle/CoreBundle/EventListener/Frontend/FrontendRoutingListener.php
@@ -16,11 +16,11 @@ namespace Pimcore\Bundle\CoreBundle\EventListener\Frontend;
 
 use Pimcore\Bundle\CoreBundle\EventListener\Traits\PimcoreContextAwareTrait;
 use Pimcore\Config;
+use Pimcore\Http\Request\Resolver\PimcoreContextResolver;
+use Pimcore\Http\Request\Resolver\SiteResolver;
 use Pimcore\Http\RequestHelper;
 use Pimcore\Model\Site;
 use Pimcore\Routing\RedirectHandler;
-use Pimcore\Service\Request\PimcoreContextResolver;
-use Pimcore\Service\Request\SiteResolver;
 use Pimcore\Tool;
 use Symfony\Component\EventDispatcher\EventSubscriberInterface;
 use Symfony\Component\HttpFoundation\RedirectResponse;
