+++ pimcore/lib/Pimcore/Bundle/CoreBundle/EventListener/Frontend/GoogleAnalyticsCodeListener.php
@@ -17,7 +17,7 @@ namespace Pimcore\Bundle\CoreBundle\EventListener\Frontend;
 use Pimcore\Bundle\CoreBundle\EventListener\Traits\PimcoreContextAwareTrait;
 use Pimcore\Bundle\CoreBundle\EventListener\Traits\ResponseInjectionTrait;
 use Pimcore\Google\Analytics as AnalyticsHelper;
-use Pimcore\Service\Request\PimcoreContextResolver;
+use Pimcore\Http\Request\Resolver\PimcoreContextResolver;
 use Symfony\Component\HttpKernel\Event\FilterResponseEvent;
 
 class GoogleAnalyticsCodeListener
