+++ pimcore/lib/Pimcore/Bundle/CoreBundle/EventListener/PimcoreContextListener.php
@@ -14,9 +14,9 @@
 
 namespace Pimcore\Bundle\CoreBundle\EventListener;
 
+use Pimcore\Http\Request\Resolver\PimcoreContextResolver;
 use Pimcore\Model\DataObject;
 use Pimcore\Model\Document;
-use Pimcore\Service\Request\PimcoreContextResolver;
 use Psr\Log\LoggerAwareInterface;
 use Psr\Log\LoggerAwareTrait;
 use Symfony\Component\EventDispatcher\EventSubscriberInterface;
