+++ pimcore/lib/Pimcore/Bundle/CoreBundle/EventListener/ResponseExceptionListener.php
@@ -17,9 +17,9 @@ namespace Pimcore\Bundle\CoreBundle\EventListener;
 use Pimcore\Bundle\CoreBundle\EventListener\Traits\PimcoreContextAwareTrait;
 use Pimcore\Config;
 use Pimcore\Http\Exception\ResponseException;
+use Pimcore\Http\Request\Resolver\PimcoreContextResolver;
 use Pimcore\Model\Document;
 use Pimcore\Model\Site;
-use Pimcore\Service\Request\PimcoreContextResolver;
 use Pimcore\Templating\Renderer\ActionRenderer;
 use Psr\Log\LoggerAwareTrait;
 use Symfony\Component\EventDispatcher\EventSubscriberInterface;
