+++ pimcore/lib/Pimcore/Document/Tag/NamingStrategy/Migration/RenderMigrationStrategy.php
@@ -23,10 +23,10 @@ use Pimcore\Bundle\AdminBundle\Session\Handler\AdminSessionHandler;
 use Pimcore\Cache;
 use Pimcore\Document\Tag\NamingStrategy\Migration\Exception\NameMappingException;
 use Pimcore\Document\Tag\NamingStrategy\Migration\Render\MigrationSubscriber;
+use Pimcore\Http\Request\Resolver\EditmodeResolver;
 use Pimcore\Model\Document;
 use Pimcore\Model\User;
 use Pimcore\Routing\Dynamic\DocumentRouteHandler;
-use Pimcore\Service\Request\EditmodeResolver;
 use Pimcore\Tool;
 use Psr\SimpleCache\CacheInterface;
 use Symfony\Component\Console\Helper\Helper;
