+++ pimcore/lib/Pimcore/Http/Request/Resolver/ResponseHeaderResolver.php
@@ -15,7 +15,7 @@ declare(strict_types=1);
  * @license    http://www.pimcore.org/license     GPLv3 and PEL
  */
 
-namespace Pimcore\Service\Request;
+namespace Pimcore\Http\Request\Resolver;
 
 use Pimcore\Controller\Configuration\ResponseHeader;
 use Symfony\Component\HttpFoundation\Request;
