+++ pimcore/lib/Pimcore/Http/Request/Resolver/TemplateResolver.php
@@ -12,7 +12,7 @@
  * @license    http://www.pimcore.org/license     GPLv3 and PEL
  */
 
-namespace Pimcore\Service\Request;
+namespace Pimcore\Http\Request\Resolver;
 
 use Pimcore\Controller\Config\ConfigNormalizer;
 use Symfony\Cmf\Bundle\RoutingBundle\Routing\DynamicRouter;
