+++ pimcore/lib/Pimcore/Http/Request/Resolver/TemplateVarsResolver.php
@@ -12,7 +12,7 @@
  * @license    http://www.pimcore.org/license     GPLv3 and PEL
  */
 
-namespace Pimcore\Service\Request;
+namespace Pimcore\Http\Request\Resolver;
 
 use Pimcore\Templating\Vars\TemplateVarsProviderInterface;
 use Symfony\Component\HttpFoundation\Request;
