+++ pimcore/lib/Pimcore/Templating/Renderer/ActionRenderer.php
@@ -15,8 +15,8 @@
 namespace Pimcore\Templating\Renderer;
 
 use Pimcore\Controller\Config\ConfigNormalizer;
+use Pimcore\Http\Request\Resolver\PimcoreContextResolver;
 use Pimcore\Model\Document;
-use Pimcore\Service\Request\PimcoreContextResolver;
 use Symfony\Bundle\FrameworkBundle\Templating\Helper\ActionsHelper;
 use Symfony\Cmf\Bundle\RoutingBundle\Routing\DynamicRouter;
 use Symfony\Component\HttpKernel\Controller\ControllerReference;
