+++ pimcore/lib/Pimcore/Tool/Glossary/Processor.php
@@ -16,12 +16,12 @@
 namespace Pimcore\Tool\Glossary;
 
 use Pimcore\Cache;
+use Pimcore\Http\Request\Resolver\DocumentResolver;
+use Pimcore\Http\Request\Resolver\EditmodeResolver;
 use Pimcore\Http\RequestHelper;
 use Pimcore\Model\Document;
 use Pimcore\Model\Glossary;
 use Pimcore\Model\Site;
-use Pimcore\Service\Request\DocumentResolver;
-use Pimcore\Service\Request\EditmodeResolver;
 
 class Processor
 {
